+++ app/src/main/java/org/simple/clinic/teleconsultlog/prescription/medicines/TeleconsultMedicinesEffectHandler.kt
@@ -6,14 +6,11 @@ import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import io.reactivex.ObservableTransformer
 import org.simple.clinic.drugs.PrescriptionRepository
-import org.simple.clinic.drugs.search.DrugFrequency
-import org.simple.clinic.drugs.selection.custom.drugfrequency.country.DrugFrequencyLabel
 import org.simple.clinic.util.scheduler.SchedulersProvider
 
 class TeleconsultMedicinesEffectHandler @AssistedInject constructor(
     private val prescriptionRepository: PrescriptionRepository,
     private val schedulersProvider: SchedulersProvider,
-    private val drugFrequencyToLabelMap: Map<DrugFrequency?, DrugFrequencyLabel>,
     @Assisted private val uiActions: TeleconsultMedicinesUiActions
 ) {
 
@@ -31,17 +28,7 @@ class TeleconsultMedicinesEffectHandler @AssistedInject constructor(
         .addConsumer(OpenDrugFrequencySheet::class.java, { uiActions.openDrugFrequencySheet(it.prescription) }, schedulersProvider.ui())
         .addConsumer(UpdateDrugDuration::class.java, { updateDrugDuration(it) }, schedulersProvider.io())
         .addConsumer(UpdateDrugFrequency::class.java, { updateDrugFrequency(it) }, schedulersProvider.io())
-        .addTransformer(LoadDrugFrequencyChoiceItems::class.java, loadDrugFrequencyChoiceItems())
         .build()
-  }
-
-  private fun loadDrugFrequencyChoiceItems(): ObservableTransformer<LoadDrugFrequencyChoiceItems, TeleconsultMedicinesEvent> {
-    return ObservableTransformer { effects ->
-      effects
-          .observeOn(schedulersProvider.io())
-          .map { drugFrequencyToLabelMap }
-          .map(::DrugFrequencyChoiceItemsLoaded)
-    }
   }
 
   private fun updateDrugFrequency(updateDrugFrequency: UpdateDrugFrequency) {