+++ app/src/main/java/org/simple/clinic/teleconsultlog/prescription/medicines/TeleconsultMedicinesEvent.kt
@@ -1,8 +1,6 @@
 package org.simple.clinic.teleconsultlog.prescription.medicines
 
 import org.simple.clinic.drugs.PrescribedDrug
-import org.simple.clinic.drugs.search.DrugFrequency
-import org.simple.clinic.drugs.selection.custom.drugfrequency.country.DrugFrequencyLabel
 import org.simple.clinic.teleconsultlog.medicinefrequency.MedicineFrequency
 import org.simple.clinic.widgets.UiEvent
 import java.time.Duration
@@ -37,7 +35,3 @@ data class DrugFrequencyChanged(
 ) : TeleconsultMedicinesEvent() {
   override val analyticsName: String = "Teleconsult Medicines:Drug Frequency Changed"
 }
-
-data class DrugFrequencyChoiceItemsLoaded(
-    val drugFrequencyToLabelMap: Map<DrugFrequency?, DrugFrequencyLabel>
-) : TeleconsultMedicinesEvent()